Preview functionality and thank you page
The preview allow you to preview a post before creating it.
Once on the preview page, you cna re-edit it, or create
To suit the needs of the how the preview works, we actually save the
job posting on preview and then call edit and update if the user wants
to re-edit. This means that it plays along nicely with paperclip,
because it needs a record to be save to access the image for preview.

To support all of this, I added a state field to jobs, with a new and
preview state. This field can also serve as a approved state once job
are approved. At some point we'll need to write a cron job to clear the
orphaned posts in preview.