Sending an email when a new job is created
Create the JobsMailer class for sending mails
Implement the new_job method to send a mail to inbox@techboard.co.za
when a new job is created.
Call the new_job method in the JobsController on Update and Create
Add the gmail sending configuration to the dev and production env files

Note: I have also setup inbox@techboard.co.za to forward mail to my
gmail so that I won't miss these mails.