Confirmation mail when job is activated
Automating what I have been doing manually up until now.
I added a action to the Jobs mailer to mail a confirmation mail to the
employer when I have activated a job. The mail is sent in the after_save
filter in the Job model if the status is active.