Fist attempt at Twitter integration
Added the tweet method to jobs. This will tweet the job title and
company as well as a bit.ly shortened link to the job.
I used the twitter and bit.ly gem, created a account on bit.ly and a
techboard job feed app on twitter, to reset up the app run the
twitter_oauth_setup.rb.

Things I would still like to do:
A rake task to call tweet on the Jobs(I'm going to do it manually
    through the console for now)